V2
Changed it so that different lines of the information tab of the game are all in one multi line variable, instead of 8 individual ones. Also the variable is global, instead of local, as it caused many problems. The quit button works, and now the "How to play" page works. All I have to do is 1. Lay out everything for the question format in the "Play", 2. Link that up to the current program to make it seamless, 3. Create parameters that lead to a GAME OVER and make a button that goes back to the home page, and 4. Further improve the program fix any bugs.