V6
Basically a troubleshooting version of the program. Did not add anything but instead tried to figure out what's going wrong in my system. I have figured out every bug other than how to display your gamerscore as a Label. The values are transferred correctly to everywhere and it is correct, but when I try to show it on the GUI it either shows 0, or nothing at all.