V7
I have added the "after()" function so that after 5 seconds of the question, the game will say game over. I have also tried to put gamescore onto it and have successfully: but the gamescore then disappears soon after with a 0. I have reason to suspect that this is due to the .after() function. I haven't had a way to redo the after() function if someone answered a question; meaning that no matter what after 5 seconds of playing the game the game will game over itself. 

2 main bugs:
 - After function will not reset after the question is answered
 - Gamescore will reset back to 0 after a short period.