V8
Nearing the end of the project:
One bug/problem remaining: When trying to use the after_cancel() command, I don't know what the after() command's specific ID is, so I need to figure that out.
Alse added 8 more questions. I plan to have a total of 20 questions in the program.